Extract class Scope from addAutoDestroyCalls and rename as AutoDestroyScope
