+++ compiler/resolution/addAutoDestroyCalls.cpp
@@ -19,7 +19,10 @@
 
 #include "addAutoDestroyCalls.h"
 
+
+
 #include "astutil.h"
+#include "AutoDestroyScope.h"
 #include "expr.h"
 #include "resolution.h"
 #include "stlUtil.h"
@@ -28,54 +31,11 @@
 
 #include <vector>
 
-/************************************* | **************************************
-*                                                                             *
-* Track the state of lexical scopes during the execution of a function.       *
-*                                                                             *
-*   1) We only track variables that have an autoDestroy flag                  *
-*                                                                             *
-*   2) The compiler introduces "formal temps" to manage formals with out and  *
-*      in-out concrete intents.  If a function has multiple returns then any  *
-*      formal temps must be handled differently from other locals.            *
-*                                                                             *
-************************************** | *************************************/
-
-class Scope
-{
-public:
-                           Scope(const Scope*     parent,
-                                 const BlockStmt* block);
-
-  void                     variableAdd(VarSymbol* var);
-
-  bool                     startingToHandleFormalTemps(const Expr* stmt) const;
-
-  void                     insertAutoDestroys(FnSymbol* fn,
-                                              Expr*     refStmt);
-
-private:
-  void                     variablesDestroy(Expr*      refStmt,
-                                            VarSymbol* excludeVar)       const;
-
-  const Scope*             mParent;
-  const BlockStmt*         mBlock;
-
-  bool                     mLocalsHandled;       // Manage function epilogue
-
-  std::vector<VarSymbol*>  mFormalTemps;         // Temps for out/inout formals
-  std::vector<VarSymbol*>  mLocals;
-};
-
-static void walkBlock(FnSymbol* fn, Scope* parent, BlockStmt* block);
-static bool isAutoDestroyedVariable(Symbol* sym);
-
-/************************************* | **************************************
-*                                                                             *
-* Entry point                                                                 *
-*                                                                             *
-************************************** | *************************************/
-
 static void cullForDefaultConstructor(FnSymbol* fn);
+
+static void walkBlock(FnSymbol*         fn,
+                      AutoDestroyScope* parent,
+                      BlockStmt*        block);
 
 void addAutoDestroyCalls() {
   forv_Vec(FnSymbol, fn, gFnSymbols) {
@@ -165,11 +125,13 @@ static LabelSymbol* findReturnLabel(FnSymbol* fn);
 static bool         isReturnLabel(const Expr*        stmt,
                                   const LabelSymbol* returnLabel);
 
-static void walkBlock(FnSymbol* fn, Scope* parent, BlockStmt* block) {
-  Scope        scope(parent, block);
+static void walkBlock(FnSymbol*         fn,
+                      AutoDestroyScope* parent,
+                      BlockStmt*        block) {
+  AutoDestroyScope scope(parent, block);
 
-  LabelSymbol* retLabel   = (parent == NULL) ? findReturnLabel(fn) : NULL;
-  bool         isDeadCode = false;
+  LabelSymbol*     retLabel   = (parent == NULL) ? findReturnLabel(fn) : NULL;
+  bool             isDeadCode = false;
 
   for_alist(stmt, block->body) {
     //
@@ -190,7 +152,7 @@ static void walkBlock(FnSymbol* fn, Scope* parent, BlockStmt* block) {
         scope.variableAdd(var);
 
       // AutoDestroy primary locals at start of function epilogue (2)
-      } else if (scope.startingToHandleFormalTemps(stmt) == true) {
+      } else if (scope.handlingFormalTemps(stmt) == true) {
         scope.insertAutoDestroys(fn, stmt);
 
       // Recurse in to a BlockStmt (or sub-classes of BlockStmt e.g. a loop)
@@ -297,209 +259,11 @@ static bool isReturnLabel(const Expr* stmt, const LabelSymbol* returnLabel) {
 
 /************************************* | **************************************
 *                                                                             *
-*                                                                             *
-*                                                                             *
-************************************** | *************************************/
-
-static VarSymbol* variableToExclude(FnSymbol*  fn, Expr* refStmt);
-static bool       isReturnStmt(const Expr* stmt);
-static BlockStmt* findBlockForTarget(GotoStmt* stmt);
-
-Scope::Scope(const Scope* parent, const BlockStmt* block) {
-  mParent        = parent;
-  mBlock         = block;
-
-  mLocalsHandled = false;
-}
-
-void Scope::variableAdd(VarSymbol* var) {
-  if (var->hasFlag(FLAG_FORMAL_TEMP) == false)
-    mLocals.push_back(var);
-  else
-    mFormalTemps.push_back(var);
-}
-
-//
-// Functions have an informal epilogue defined by code that
-//
-//   1) appears after a common "return label" (if present)
-//   2) copies values to out/in-out formals
-//
-// We must      destroy the primaries before (1).
-// We choose to destroy the primaries before (2).
-//
-// This code detects the start of (2)
-//
-bool Scope::startingToHandleFormalTemps(const Expr* stmt) const {
-  bool retval = false;
-
-  if (mLocalsHandled == false) {
-    if (const CallExpr* call = toConstCallExpr(stmt)) {
-      if (FnSymbol* fn = call->resolvedFunction()) {
-        if (fn->hasFlag(FLAG_ASSIGNOP) == true && call->numActuals() == 2) {
-          SymExpr* lhs = toSymExpr(call->get(1));
-          SymExpr* rhs = toSymExpr(call->get(2));
-
-          if (lhs                                      != NULL &&
-              rhs                                      != NULL &&
-              isArgSymbol(lhs->symbol())               == true &&
-              rhs->symbol()->hasFlag(FLAG_FORMAL_TEMP) == true) {
-            retval = true;
-          }
-        }
-      }
-    }
-  }
-
-  return retval;
-}
-
-// If the refStmt is a goto then we need to recurse
-// to the block that contains the target of the goto
-void Scope::insertAutoDestroys(FnSymbol* fn, Expr* refStmt) {
-  GotoStmt*    gotoStmt   = toGotoStmt(refStmt);
-  bool         recurse    = (gotoStmt != NULL) ? true : false;
-  BlockStmt*   forTarget  = findBlockForTarget(gotoStmt);
-  VarSymbol*   excludeVar = variableToExclude(fn, refStmt);
-  const Scope* scope      = this;
-
-  while (scope != NULL && scope->mBlock != forTarget) {
-    scope->variablesDestroy(refStmt, excludeVar);
-
-    scope = (recurse == true) ? scope->mParent : NULL;
-  }
-
-  mLocalsHandled = true;
-}
-
-void Scope::variablesDestroy(Expr* refStmt, VarSymbol* excludeVar) const {
-  // Handle the primary locals
-  if (mLocalsHandled == false) {
-    bool   insertAfter = false;
-    size_t count       = mLocals.size();
-
-    // If this is a simple nested block, insert after the final stmt
-    // Do not get tricked by sequences of unreachable code
-    if (refStmt->next == NULL) {
-      if (mParent != NULL && isGotoStmt(refStmt) == false) {
-        insertAfter = true;
-      }
-    }
-
-    for (size_t i = 1; i <= count; i++) {
-      VarSymbol* var = mLocals[count - i];
-
-      if (var != excludeVar) {
-        if (FnSymbol* autoDestroyFn = autoDestroyMap.get(var->type)) {
-          SET_LINENO(var);
-
-          INT_ASSERT(autoDestroyFn->hasFlag(FLAG_AUTO_DESTROY_FN));
-
-          CallExpr* autoDestroy = new CallExpr(autoDestroyFn, var);
-
-          if (insertAfter == true)
-            refStmt->insertAfter (autoDestroy);
-          else
-            refStmt->insertBefore(autoDestroy);
-        }
-      }
-    }
-  }
-
-  // Handle the formal temps
-  if (isReturnStmt(refStmt) == true) {
-    size_t count = mFormalTemps.size();
-
-    for (size_t i = 1; i <= count; i++) {
-      VarSymbol* var = mFormalTemps[count - i];
-
-      if (FnSymbol* autoDestroyFn = autoDestroyMap.get(var->type)) {
-        SET_LINENO(var);
-
-        refStmt->insertBefore(new CallExpr(autoDestroyFn, var));
-      }
-    }
-  }
-}
-
-// Walk backwards from the current statement to determine if a sequence of
-// moves have copied a variable that is marked for auto destruction in to
-// the dedicated return-temp within the current scope.
-//
-// Note that the value we are concerned about may be copied in to one or
-// more temporary variables between being copied to the return temp.
-static VarSymbol* variableToExclude(FnSymbol*  fn, Expr* refStmt) {
-  VarSymbol* retVar = toVarSymbol(fn->getReturnSymbol());
-  VarSymbol* retval = NULL;
-
-  if (retVar != NULL) {
-    if (isUserDefinedRecord(retVar)    == true ||
-        fn->hasFlag(FLAG_INIT_COPY_FN) == true) {
-      VarSymbol* needle = retVar;
-      Expr*      expr   = refStmt;
-
-      // Walk backwards looking for the variable that is being returned
-      while (retval == NULL && expr != NULL && needle != NULL) {
-        if (CallExpr* move = toCallExpr(expr)) {
-          if (move->isPrimitive(PRIM_MOVE) == true) {
-            SymExpr*   lhs    = toSymExpr(move->get(1));
-            VarSymbol* lhsVar = toVarSymbol(lhs->symbol());
-
-            if (needle == lhsVar) {
-              SymExpr*   rhs    = toSymExpr(move->get(2));
-              VarSymbol* rhsVar = (rhs != NULL) ? toVarSymbol(rhs->symbol()) : NULL;
-
-              if (isAutoDestroyedVariable(rhsVar) == true)
-                retval = rhsVar;
-              else
-                needle = rhsVar;
-            }
-          }
-        }
-
-        expr = expr->prev;
-      }
-    }
-  }
-
-  return retval;
-}
-
-static bool isReturnStmt(const Expr* stmt) {
-  bool retval = false;
-
-  if (const CallExpr* expr = toConstCallExpr(stmt))
-    retval = expr->isPrimitive(PRIM_RETURN);
-
-  return retval;
-}
-
-// Find the block stmt that encloses the target of this gotoStmt
-static BlockStmt* findBlockForTarget(GotoStmt* stmt) {
-  BlockStmt* retval = NULL;
-
-  if (stmt != NULL && stmt->isGotoReturn() == false) {
-    SymExpr*   labelSymExpr = toSymExpr(stmt->label);
-    Expr*      ptr          = labelSymExpr->symbol()->defPoint;
-
-    while (ptr != NULL && isBlockStmt(ptr) == false)
-      ptr = ptr->parentExpr;
-
-    retval = toBlockStmt(ptr);
-
-    INT_ASSERT(retval);
-  }
-
-  return retval;
-}
-
-/************************************* | **************************************
-*                                                                             *
-* Common utilities                                                            *
+* A shared utilty function for resolution.                                    *
 *                                                                             *
 ************************************** | *************************************/
 
-static bool isAutoDestroyedVariable(Symbol* sym) {
+bool isAutoDestroyedVariable(Symbol* sym) {
   bool retval = false;
 
   if (VarSymbol* var = toVarSymbol(sym)) {
